chore(sdk): remove aws-cdk-lib from dependencies
