+++ libs/awscdk/test/on-deploy.test.ts
@@ -2,7 +2,8 @@ import { Match, Template } from "aws-cdk-lib/assertions";
 import { expect, test } from "vitest";
 import { cloud, simulator } from "@winglang/sdk";
 import * as awscdk from "../src";
-import { mkdtemp, awscdkSanitize } from "@winglang/sdk/test/util";
+import { mkdtemp } from "@winglang/sdk/test/util";
+import { awscdkSanitize } from "./util";
 
 const CDK_APP_OPTS = {
   stackName: "my-project",
