+++ libs/awscdk/test/util.ts
@@ -1,3 +1,5 @@
+import { Template } from "aws-cdk-lib/assertions";
+
 /**
  * Sanitize the text of a code bundle to remove path references that are system-specific.
  */
@@ -10,4 +12,23 @@ export function sanitizeCode(code: string): string {
   }
 
   return removeAbsolutePaths(code);
+}
+
+export function awscdkSanitize(template: Template): any {
+  let json = template.toJSON();
+
+  let jsonString = JSON.stringify(json, (key, value) => {
+    if (key === "S3Key" && value.endsWith(".zip")) {
+      return "<S3Key>";
+    }
+
+    return value;
+  });
+
+  jsonString = jsonString.replace(
+    /CurrentVersion.+?"/g,
+    'CurrentVersion<GUID>"'
+  );
+
+  return JSON.parse(jsonString);
 }
